Fix UUID consistency in patient service/controller layers and related compilation errors

+++ backend/src/main/java/com/shifa/service/doctor/DashboardStatsService.java
@@ -34,7 +34,7 @@ public class DashboardStatsService {
                 .monthNewPatients((int)visitRepository.countNewPatientsByDoctorSince(doctorId, monthStart))
                 .pendingAiProcessing((int)visitRepository.countByDoctorUserIdAndStatus(doctorId, VisitStatus.NOTES_TAKEN))
                 .pendingSendToPatient((int)visitRepository.countByDoctorUserIdAndStatus(doctorId, VisitStatus.REVIEWED))
-                .failedNotifications(notificationRepository.countFailedNotificationsSince(weekAgo))
+                .failedNotifications((int)notificationRepository.countFailedNotificationsSince(weekAgo))
                 .totalPatients((int)visitRepository.countDistinctPatientsByDoctorUserId(doctorId))
                 .totalVisits((int)visitRepository.countByDoctorUserIdAndDeletedFalse(doctorId))
                 .whatsappReadRate(calculateReadRate(doctorId))
